Mascot_Git/requirements.txt
Flask
flask-cors
python-dotenv
google-genai
pydantic-ai
nest_asyncio